Reparent odp-core submodule
odp-core is moved down the build project's 'submodule tree' into odp-server. This solves the immediate problem of linking odp-core in for the server CI tests. But conceptually it makes sense for the client projects (odp-admin and odp-web) to depend on the odp-core of the server with which they are working.