Update client API with client-collection relationship change
list/get route authorizations now enforce only the relevant scope access; they do not consider whether returned client objects are collection-specific even when the authorization is. This is fine because a user with a collection-specific role would not in any case be able to do anything with a collection id for which they do not have access.

Further, we will in the near future be changing collection ids to be opaque uuids, so there will be no problem with inappropriate visibility thereof.