Use UUIDs for collection and provider primary keys
And use the new 'abbr' column for human-readable short ids.

The provider, collection and record tables will need to be truncated before running the SQL migrations for these changes - which is a good reason for getting this done before we're in production. The ODP v1-to-v2 migration will need to be modified accordingly and re-run.