Tidy up identity exception handling
HydraAdminError is not raised by anything (it was by the old Hydra admin client which has been replaced by the Hydra SDK). Hydra API errors will (still) emerge in the UI as internal server errors which is fine - the removed Hydra exception handling code was effectively doing the same thing.