Create indexes for the catalog record search API
Note that coalesce(temporal_end, temporal_start) in the API seems to be resistant to indexing - I could not prevent the execution plan doing a filter instead of using an index, even when an index was created which exactly matched coalesce(temporal_end, temporal_start). In any case, the use of coalesce suggests that temporal_end can rather be set to temporal_start at publishing time if null - which should allow the simple index on temporal_end to just work.