Implement generic keyword tag schema
This replaces the SDG-specific tag schema and provides a generic capability for tagging records with keywords from controlled vocabularies. Ultimately, this may prove useful in terms of managing keywords that are expressed in a record's metadata.

This approach may be extended to project/infrastructure collection tags, which will enable us to do away with the somewhat clunky (and no longer necessary, with the new tag-vocabulary link) "vocabulary" JSON Schema extension keyword.