Check out submodule branches based on odp-server branch
For PR's this will require matching branch names to exist in odp-core and odp-ui. This makes sense for odp-core, which is usually developed in conjunction with odp-server. However, it seems contrived to create an odp-ui branch just for server dev. In future, UI dependencies might be removed from the server codebase (by breaking out odp-identity).