Introduce archive_resource many-to-many relation
This separates the problems of resource definition and resource archival. It allows for a resource to be uploaded without prior knowledge of where it may be stored, and for an archival process to run independently and without modifying resource definitions themselves. It also allows mirroring of resources across archives, which may assist in upload, migration and download use cases.