Remove collection-specific auth from client and role APIs
The present code caters to the possibility of client/role admin access being granted to collection-specific clients or roles, and disallowing changes to any client/role involving collections to which a client or user does not have access.

While this is strictly correct, it is unnecessary. Collection-specificity in clients and roles was designed for the real use case of limiting access to collections and records, while clients and roles are configuration objects that are managed by administrators.

With the pending development of provider-specific auth for package access, it does not make sense to further cater to unlikely authorization scenarios for client/role config. In any case, authorization responses will likely be tweaked such that collection-specific and provider-specific restrictions target only applicable scopes (e.g. collection.*, record.*, package.*, etc).