Remove dependence of user info on client
Previously, any role which would be ineffective in the ODP for the client from which user info was requested (by virtue of being constrained to a different set of collections as the client) was excluded from the list of roles returned in the user info. This contrived edge case logic, while not incorrect, misses three points:

- Some returned roles might be ineffective for the client regardless of being collection-specific or not, because the client in question might not care about ODP roles at all.
- A client might choose to use or display role information even for roles which are ineffective for it in the ODP.
- Providing visibility of an ineffective role does not magically endow it with any kind of access; ODP authorization logic will in any case prevent its use from that client.

In principle, the revised user info response simply gives the state of an ODP user.