Revise ERD for ODP v3
Besides improved labelling and layout, this commit makes several important changes to the information architecture of the ODP.

The provider-collection one-to-many relationship is dropped. As discussed in #23, provider objects facilitate grouping and authorization for data submission (package management), while collections facilitate grouping and authorization for data curation (metadata management). They are not directly related. This closes #23.

The recently introduced provider-role one-to-many relationship is dropped, and replaced with a provider-user many-to-many. In discussion with the curation team we agreed that, in general, access to packages should be partitioned on an individual user basis - not grouped by institution or department. This means that the number of providers will grow significantly - and having to configure per-provider roles would not be scalable. A provider-user association table will support the partitioning of package/resource access by user, while also allowing the flexibility for multiple users to form a provider group. Note that existing provider info will need to be migrated to institution tags or some such.

The client-collection many-to-many relationship is dropped, and replaced with an optional provider-client one-to-many. Automated data providers (such as the SAEON Observations Database) will now create packages rather than records, and access will be constrained by provider rather than collection. There is no need for collection-specific clients because the ODP will no longer support automated creation/update of records by external clients. Collection-specific roles still allow for collection-constrained access to curation functions via UI applications, where needed.

A package-tag many-to-many relationship is added. This will allow data providers with the applicable permission to tag packages with information (e.g. a project keyword) to support automatic harvesting and publication of package metadata. This will be particularly useful for automated data providers.

A schema-package one-to-many relationship is added. Metadata from external users and clients will now arrive via packages, so it makes sense to provide validation and feedback for such. Ultimately, preserving submission metadata in packages may allow room to relationally model the metadata that we currently store as JSON in record objects.

Note that scope-tag and scope-vocabulary relationships have been deleted from the diagram (and a possible scope-archive has not been added) to reduce clutter. These are in any case auth implementation details and diagramming them detracts from the fact that scopes apply to everything.