Normalize usage of id:value dicts in API output models
Previously we had dicts of collection key:id and record doi:id, but this approach does not work for e.g. user name:id, since name is not necessarily unique. We can easily adapt the UI code that makes use of these dicts to expect the ids to be in keys rather than in values.