Implement permissions for user access to provider-constrainable scopes
Note that in the user permissions calculation, we no longer remove a (constrainable) scope from the result if its applicability is the empty set. This was being done inconsistently and unnecessarily: API routes do in any case disallow access for a constrainable scope with an empty object set, and UI clients can in any case introspect these object sets to enable/disable buttons (for example) on a per-object basis.

The auth tests now parameterize the batch sizes of collections associated with collection-specific test roles and providers associated with the test user. This gives us better coverage of set combinations including cases where the applicable sets are empty. The increased number of invocations additionally tests a greater variety of scope sample sets within a given test suite run.