Audit related objects
An audited entity must track changes to many-to-many relations if it is the 'master' of the relationship (related objects are managed via its own API). We're doing this already for provider audit (related users) and user/identity audit (related roles). Now adding DB support for auditing changes to packages' resources and records' packages.