Execute auth dependency first
If a user/client does not have the requisite API scope, they should get a 403 even if their input is invalid and they would otherwise get some other kind of error. Negative tests for the suggest keyword API have been updated to parameterize scopes as for the positive test, to check that a 403 is returned in unauthorized cases in preference to a 409, 422, etc.