Rename keyword primary key to `id`
For consistency with ODP primary keys, in general. Note also that the keyword id is not necessarily what will be displayed in a UI or rendered to a metadata record. For example, an institution keyword id might be 'Institution:SAEON', while 'SAEON' would be the effective value of the keyword.