Update API usage of keyword ids and parent ids
The keyword regular expression has been removed. The only keyword format requirement - which is enforced in the DB layer - is that a keyword is formed by a non-empty suffix to its parent keyword. This allows us the freedom to structure keywords in whatever way is most appropriate for a vocabulary.