Add package contributor tag and schema
The contributor schema includes a speculative affiliations array, items of which may be either a "$ref" to the institution schema (accommodating unlisted institutions) or a "keyword" reference to the institution schema. The "keyword" keyword has not yet been implemented, but the idea is that this will behave like an enum such that the instance value matches an entry in the specified ODP keyword vocabulary.

The contributor schema also includes speculative translations to DataCite creators and contributors, with the affiliation property delegated to an institution schema translation via a "$ref". Further development is needed in the jschon-translation library to enable this to work.