Add path to package_resource association table
This represents a relative path to a resource within a package and provides a way to prevent duplication of resources within packages. This is distinct from the archival path for a resource on archive_resource.

Note that the ORM relationship configuration on package has changed such that resource instances can no longer simply be assigned to a package to create a package_resource. The additional association attributes require package_resource instances to be created explicitly.

The notes column is also dropped from package; a user-cardinality package notes tag would be more sensible.