Add package key and generalize resource hash
The idea with package key is that it will be a sanitized form of the package title, and will form part of the archival path for a resource upload, as in {archive.url}/{provider.key}/{package.key}/{file_path}, instead of using the package uuid in that path position. This will make it easier for curators/admins to see at a glance what package a folder relates to when viewing it in Nextcloud. The API will need to consider how to deal with package title changes. It doesn't really matter from an archive/package/resource management perspective, but it might be confusing for curators/admins if a package's files are split over multiple directories.

Resource title is also made nullable. This is more sensible when we start looking at multiple file uploads and really it is unnecessary. We do however add a check such that title must be given if filename is not (relevant for externally managed datasets e.g. Obs DB).