Fix each resource to a single package
The package_resource many-to-many relation is dropped. The ability to associate a resource with multiple packages was a fanciful idea with no basis in reality. Resubmissions, for example, are treated by curators as standalone packages, even if some of the contents remain unchanged. In principle, every package should be standalone and self-contained in any case, and allowing a resource to be shared amongst packages (not to mention providers) would cause a great deal of complexity and ambiguity.

Fixing a resource to a single package ensures unambiguous ownership (provider-package-resource) and location (provider/package/folder/file).

Besides, the symmetry :magic_wand: