Reinstate vocabulary; rename resource to file
The current keyword infrastructure (DB and API) is overcomplicated; it tries to be too much. There is contrived code for dealing with top-level keywords (which implicitly represent vocabularies) and chunky composition of keyword ids (prefixed with top-level/vocab id - not 'normal' in any sense). Custom schemas for nested vocabs? Should be an altogether different vocab.

Our primary use case for hierarchical keywords - SDGs - can be effectively implemented with a single vocabulary schema that supports multiple levels, and parent-referencing keywords which each reference the SDG vocab. The vocab id is part of the keyword primary key. No problem with same keywords in different vocabs. No need for child vocab schemas.

We also rename resource to file. Because in Unix, everything is a file! Also, 'resource' could mean just about anything, whereas 'file' implies a reliably referenceable, packageable instance of data. Of course, a file could be an external web folder, but to the ODP, it's still a file that can be published and accessed via the generic catalog-record-package-file-archive pipeline.