Revert 'file' to 'resource'
This reverts the diagram-only rename of 'resource' to 'file' made as part of 08655c5ddaa11c60267f2b50be8a6302966501bc

As mentioned in the previous ERD commit, a resource can be a (web) folder, and we already split resource location into folder and filename parts - with filename blank in the case of a folder. In concordance with the generic output record, let's stick with the generic input resource. This avoids misinterpreting a resource as necessarily representing a file-like object, and also avoids naming conflicts with actual file-like objects in code.