Rework vocabulary and keyword DB models
Ref 08655c5ddaa11c60267f2b50be8a6302966501bc

Vocabulary had been deprecated in favour of a self-referencing keyword, which was implicitly a vocabulary if it had child keywords. Vocabulary is now restored; keyword replaces vocabulary_term; keywords can be hierarchical but every keyword in a hierarchy belongs explicitly to a vocabulary. [Can a keyword hierarchy cross vocabularies? Maybe - further analysis needed.]

A keyword's schema is now defined by its vocabulary; the vocabulary schema can cater to variations at different levels in a keyword hierarchy.

The contrived formation of keyword id by suffixing parent id is binned. This was done based on how SDG keywords are formed (goal.target.indicator). In general, there's no reason why sub-keywords have to be prefixed with their parents. Now, we only need to ensure uniqueness of a keyword within its vocabulary.

The keyword itself is now no longer the primary key. We use an autoinc id field instead, and a key field for the keyword. This caters better to potentially variable and/or user-supplied keywords such as institution names or acronyms. The autoinc id serves as a persistent keyword code, while the keyword itself can be changed as needed without adverse effects.