Rework static vocabulary initialization
Static keyword definitions are moved out of systemdata YAML files and into JSON files under odp-core. This gives static ODP vocabularies a similar level of importance and accessibility as the ODP JSON schemas.

The SDG keyword schema is simplified. For keyword objects in general, the key column is included in the data (vocab-specific keyword format can thus be validated), while id is not (since ids are autogenerated). Level indicators and parent refs are removed. Keyword hierarchy is determined (for static vocabs) by the nested structure in the vocab JSON, and controlled by the self-relationship on the keyword table.