Drop package title
Title was made a package attribute instead of a tag so that it could be used to generate a package key, which is a human-readable value that can be used as the package folder name on storage servers. But this approach is flawed, and it shows in the ugly code that handles changes to package title: if the title changes, then we update the package key, but not if the package already has any associated resources because then we might end up with package files split across multiple folders. :facepalm:

Now instead we format package key using provider key, date and a counter, so the key is set once only. And title can become a package tag that encodes its own translations and can be fed into the translation machinery like all other package metadata.