Replace existing objects
It's a PUT. This is expected.

Also, combine folder and filename into path across the board. This simplifies path integrity checks and validation, and parts can in any case be easily inferred by the client. If in future we need to distinguish file and folder resource types, this can be done explicitly with an enum on resource, instead of implicitly based on folder/filename values.